changed to float, DCT parallel

+++ compression.cpp
@@ -8,15 +8,15 @@ using namespace std;
 using namespace cv;
 
 
-vector<vector<double>> dct(int N) {
-    vector<vector<double>> T(N, vector<double>(N, 0.0));
+vector<vector<float>> dct(int N) {
+    vector<vector<float>> T(N, vector<float>(N, 0.0));
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             if (i == 0){
                 T[i][j] = 1/sqrt(N);
             }
             else{
-                double tmp = ((2*j+1)*i*M_PI)/(2*N);
+                float tmp = ((2*j+1)*i*M_PI)/(2*N);
                 T[i][j] = sqrt(2)/sqrt(N) * cos(tmp);
             }
         }
@@ -24,8 +24,8 @@ vector<vector<double>> dct(int N) {
     return T;
 }
 
-vector<vector<double>> transpose(vector<vector<double>> &matrix , int N){
-    vector<vector<double>> mat_tran (N, vector<double>(N, 0.0));
+vector<vector<float>> transpose(vector<vector<float>> &matrix , int N){
+    vector<vector<float>> mat_tran (N, vector<float>(N, 0.0));
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             mat_tran[i][j] = matrix[j][i];
@@ -35,8 +35,8 @@ vector<vector<double>> transpose(vector<vector<double>> &matrix , int N){
 }
 
 
-vector<vector<double>> mat_mul(vector<vector<double>> &matrix1, vector<vector<double>> &matrix2, int N){
-    vector<vector<double>> res(N, vector<double>(N,0));
+vector<vector<float>> mat_mul(vector<vector<float>> &matrix1, vector<vector<float>> &matrix2, int N){
+    vector<vector<float>> res(N, vector<float>(N,0));
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             for (int k = 0; k < N; k ++){
@@ -47,13 +47,13 @@ vector<vector<double>> mat_mul(vector<vector<double>> &matrix1, vector<vector<do
     return res;
 }
 
-vector<vector<double>> muls(vector<vector<double>> &M,vector<vector<double>> &T,vector<vector<double>> &T_tran,int N){
-    vector<vector<double>> tmp = mat_mul(T,M,N);
-    vector<vector<double>> D = mat_mul(tmp,T_tran,N);
+vector<vector<float>> muls(vector<vector<float>> &M,vector<vector<float>> &T,vector<vector<float>> &T_tran,int N){
+    vector<vector<float>> tmp = mat_mul(T,M,N);
+    vector<vector<float>> D = mat_mul(tmp,T_tran,N);
     return D;
 }
 
-void print(vector<vector<double>> &matrix, int N){
+void print(vector<vector<float>> &matrix, int N){
     for (int i = 0; i < N; i++) {
         for (int j = 0; j < N; j++) {
             cout << matrix[i][j] << "\t"; // Print each element followed by a tab
@@ -62,8 +62,8 @@ void print(vector<vector<double>> &matrix, int N){
     }
 }
 
-vector<vector<double>> compress(vector<vector<double>> &Q, vector<vector<double>> &D, int N){
-    vector<vector<double>> C (N, vector<double>(N, 0.0));
+vector<vector<float>> compress(vector<vector<float>> &Q, vector<vector<float>> &D, int N){
+    vector<vector<float>> C (N, vector<float>(N, 0.0));
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             C[i][j] = round(D[i][j]/Q[i][j]);
@@ -72,8 +72,8 @@ vector<vector<double>> compress(vector<vector<double>> &Q, vector<vector<double>
     return C;
 }
 
-vector<vector<double>> decompress(vector<vector<double>> &Q, vector<vector<double>> &C, int N){
-    vector<vector<double>> R (N, vector<double>(N, 0.0));
+vector<vector<float>> decompress(vector<vector<float>> &Q, vector<vector<float>> &C, int N){
+    vector<vector<float>> R (N, vector<float>(N, 0.0));
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             R[i][j] = round(Q[i][j] * C[i][j]);
@@ -81,14 +81,14 @@ vector<vector<double>> decompress(vector<vector<double>> &Q, vector<vector<doubl
     }
     return R;
 }
-void sub128(vector<vector<double>> &M, int N){
+void sub128(vector<vector<float>> &M, int N){
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             M[i][j] -= 128;
         }
     }
 }
-void add128(vector<vector<double>> &M, int N){
+void add128(vector<vector<float>> &M, int N){
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             M[i][j] += 128;
@@ -96,7 +96,7 @@ void add128(vector<vector<double>> &M, int N){
     }
 }
 
-void quant(vector<vector<double>> &M, int N, int q){
+void quant(vector<vector<float>> &M, int N, int q){
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             M[i][j] *= q;
@@ -104,7 +104,7 @@ void quant(vector<vector<double>> &M, int N, int q){
     }
 }
 
-void mat_round(vector<vector<double>> &M, int N){
+void mat_round(vector<vector<float>> &M, int N){
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
             M[i][j] = round(M[i][j]);
@@ -112,10 +112,10 @@ void mat_round(vector<vector<double>> &M, int N){
     }
 }
 
-vector<vector<double>> image_compression(vector<vector<double>> &M , int N){
-    vector<vector<double>> T =  dct(N);
-    vector<vector<double>> T_tran = transpose(T,N);
-    vector<vector<double>> Q = {
+vector<vector<float>> image_compression(vector<vector<float>> &M , int N){
+    vector<vector<float>> T =  dct(N);
+    vector<vector<float>> T_tran = transpose(T,N);
+    vector<vector<float>> Q = {
         {16, 11, 10, 16, 24, 40, 51, 61},
         {12, 12, 14, 19, 26, 58, 60, 55},
         {14, 13, 16, 24, 40, 57, 69, 56},
@@ -127,12 +127,12 @@ vector<vector<double>> image_compression(vector<vector<double>> &M , int N){
     };
     quant(Q,N,5);
     sub128(M,N);
-    vector<vector<double>> D = muls(M,T,T_tran,N);
-    vector<vector<double>> C = compress(Q,D,N);
+    vector<vector<float>> D = muls(M,T,T_tran,N);
+    vector<vector<float>> C = compress(Q,D,N);
     // we need to decode C
-    vector<vector<double>> R = decompress(Q,C,N);
+    vector<vector<float>> R = decompress(Q,C,N);
 
-    vector<vector<double>> res = muls(R,T_tran,T,N);
+    vector<vector<float>> res = muls(R,T_tran,T,N);
     mat_round(res,N);
     add128(res,N);
 
@@ -159,7 +159,7 @@ void read_image(string filename){
     int block_size = 8;
 
     // Vector to store 8x8 regions as vectors of vectors
-    vector<vector<vector<double>>> blockVectors;
+    vector<vector<vector<float>>> blockVectors;
 
     // Loop through the image and convert each 8x8 block to vector of vectors
     for (int y = 0; y < rows; y += block_size) {
@@ -169,12 +169,12 @@ void read_image(string filename){
             Rect roi(x, y, min(block_size,cols-x), min(block_size,rows-y));
             Mat region = grayImage(roi);
             // cout<<region.rows<< " "<<region.cols<<endl;
-            // Convert Mat region to a vector<vector<double>>
-            vector<vector<double>> blockVector;
+            // Convert Mat region to a vector<vector<float>>
+            vector<vector<float>> blockVector;
             for (int i = 0; i < region.rows; ++i) {
-                vector<double> rowVector;
+                vector<float> rowVector;
                 for (int j = 0; j < region.cols; ++j) {
-                    rowVector.push_back(static_cast<double>(region.at<uchar>(i, j)));
+                    rowVector.push_back(static_cast<float>(region.at<uchar>(i, j)));
                 }
                 blockVector.push_back(rowVector);
             }
@@ -183,7 +183,7 @@ void read_image(string filename){
                 blockVectors.push_back(blockVector);
                 continue;
             }
-            vector<vector<double>> updated_blockVector = image_compression(blockVector,block_size);
+            vector<vector<float>> updated_blockVector = image_compression(blockVector,block_size);
             // return;
             // Store the block vector
             blockVectors.push_back(updated_blockVector);
@@ -195,7 +195,7 @@ void read_image(string filename){
     for (int y = 0; y < rows; y += block_size) {
         for (int x = 0; x < cols; x += block_size) {
             // Get the block vector
-            vector<vector<double>> blockVector = blockVectors[index++];
+            vector<vector<float>> blockVector = blockVectors[index++];
 
             // Convert the block vector back to Mat format
             for (int i = 0; i < (int)blockVector.size(); ++i) {
